pyproject.toml
[project]
name = "agentkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-protocol>=0.1.7",
    "aio-pika>=9.5.5",
    "langchain>=0.3.25",
    "langchain-core>=0.3.65",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.1",
    "python-socketio[client]>=5.13.0",
]

[dependency-groups]
dev = [
    "langchain>=0.3.26",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "langgraph-checkpoint-postgres>=2.0.21",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
]
